Final iteration of camera switch
The final iteration of camera switching. Its a simple setup. Add the collider volume blueprint to the scene and then add a camera to the scene. Position the camera where you want it and connect it to the collider in in the setting under "Cameretogoto". In our case, multiple cameras in the terrarium, the first camera needs to cover every other collider, this will decide whether you are in or out of the terrarium. This will be the main cover collider and do not forget to tick "Cover collider" box in the settings once its added to the scene.